tests: rework test_retention_with_node_failures
This test relied on steady rate from the producer,
which VerifiableProducer does not do well.  Replace
this with KgoVerifier that uses golang's x/rate crate
for a more dependable token bucket rate limit.

Relocate the test to cloud_retention_test, as it is no
longer an EndToEndTest, and makes more sense there anyway.

Fixes https://github.com/redpanda-data/redpanda/issues/8577
